Allow for more permissive CORS behaviour in non production environments
This PR Is more of a suggestion / discussion point.

When using the default RemixAppServer as a development server I hit CORS issues on the statically served files from the `public` folder, where there is no way to set more permissive CORS headers. This change would allows RemixAppServer to continue to be used in development in my case, where I'm accessing the served page in and `<iframe sandbox="allow-scripts" />'.

Should we have permissive CORS headers added in development by default?